index dataset service almost complete

+++ src/indexes/ElasticClient.test.ts
@@ -1,13 +1,18 @@
+import { nanoid } from "nanoid"
 import { ElasticClient } from "./ElasticClient"
 
-describe.skip("ElasticClient test",() =>{
+describe("ElasticClient test",() =>{
 
     const elasticTestUrl = process.env.TEST_ES_INDEX_URL
+
+    if (!elasticTestUrl) {
+        throw new Error("Cannot run test no process.env.TEST_ES_INDEX_URL specified")
+    }
 
     const indexClient = new ElasticClient(elasticTestUrl)
     const testIndex = "test-index"
     
-    test.skip("It should delete document in index by query", async () => {
+    test("It should delete document in index by query", async () => {
 
         await indexClient.deleteIndex({
             index: testIndex
@@ -34,6 +39,7 @@ describe.skip("ElasticClient test",() =>{
         })
 
         const johnDoe = res.body.hits.hits[0]._source
+
 
         expect(johnDoe).toBeDefined()
         expect(johnDoe.name).toBe("John")
@@ -64,5 +70,43 @@ describe.skip("ElasticClient test",() =>{
 
         await indexClient.deleteIndex({ index: testIndex})
                 
+    })
+
+    test("It should bulk load document into index", async () => {
+
+        const index = "bulk-" + nanoid().toLowerCase()
+
+
+        const docs = [
+            {
+                name: "Pluto",
+                age: 20,
+                id: 1
+            },
+            {
+                name: "Minnie",
+                age: 30,
+            }
+        ]
+
+        await indexClient.loadBulkDocuments(index, docs)
+
+        const res = await indexClient.searchDocuments(index, {
+            query : {
+                match_all : {}
+            }
+        })
+
+        const uploaded = res.body.hits.hits
+        expect(uploaded).toBeDefined()
+
+        const pluto = res.body.hits.hits[0]._source
+
+        expect(pluto.age).toBe(20)
+        expect(pluto.name).toBe("Pluto")
+
+
+        await indexClient.deleteIndex({ index: index })
+
     })
 })
